Refactor BarAreaCharts component for improved data handling
