+++ app/components/Layout/Header/NavbarLogo.tsx
@@ -1,11 +1,11 @@
 import { ToggleHamburgerHorizontalIcon } from "@/app/icons"
-import { selectThemeConfig, themeConfigSlice } from '@/lib/redux/slices/themeConfigSlice';
+import { toggleSidebar } from "@/lib/features/themeConfig/themeConfigSlice";
+import { useAppDispatch } from "@/lib/hooks";
 import Image from "next/image"
 import Link from "next/link"
-import { useDispatch } from "react-redux";
 
 const NavbarLogo = () => {
-    const dispatch = useDispatch();
+    const dispatch = useAppDispatch();
     return (
         <div className="horizontal-logo flex items-center justify-between ltr:mr-2 rtl:ml-2 lg:hidden">
             <Link href="/" className="main-logo flex shrink-0 items-center">
@@ -15,7 +15,7 @@ const NavbarLogo = () => {
             <button
                 type="button"
                 className="collapse-icon  flex flex-none rounded-full bg-white-light/40 p-2 hover:bg-white-light/90 hover:text-primary ltr:ml-2 rtl:mr-2 dark:bg-dark/40 dark:text-[#d0d2d6] dark:hover:bg-dark/60 dark:hover:text-primary lg:hidden"
-                onClick={() => dispatch(themeConfigSlice.actions.toggleSidebar())}
+                onClick={() => dispatch(toggleSidebar())}
             >
                 <ToggleHamburgerHorizontalIcon />
             </button>
