+++ container/AppContainer.tsx
@@ -1,27 +1,27 @@
 "use client";
-import { selectThemeConfig, themeConfigSlice, } from "@/lib/redux/slices/themeConfigSlice";
 import { PropsWithChildren, useEffect } from "react";
 // import { useTranslation } from 'react-i18next';
-import { useDispatch, useSelector } from 'react-redux';
 // import { IRootState } from './store';
 // import { toggleRTL, toggleTheme, toggleLocale, toggleMenu, toggleLayout, toggleAnimation, toggleNavbar, toggleSemidark } from './store/themeConfigSlice';
 import { MantineProvider } from '@mantine/core';
+import { useAppDispatch, useAppSelector } from "@/lib/hooks";
+import { selectThemeConfig, toggleAnimation, toggleLayout, toggleMenu, toggleNavbar, toggleRTL, toggleSemidark, toggleTheme } from "@/lib/features/themeConfig/themeConfigSlice";
 
 function AppContainer({ children }: PropsWithChildren) {
 
-    const themeConfig = useSelector(selectThemeConfig)
+    const themeConfig = useAppSelector(selectThemeConfig)
     // const themeConfig = useSelector((state: IRootState) => state.themeConfig);
-    const dispatch = useDispatch();
+    const dispatch = useAppDispatch();
     // const { i18n } = useTranslation();
 
     useEffect(() => {
-        dispatch(themeConfigSlice.actions.toggleTheme(localStorage.getItem('theme') || themeConfig.theme));
-        dispatch(themeConfigSlice.actions.toggleMenu(localStorage.getItem('menu') || themeConfig.menu));
-        dispatch(themeConfigSlice.actions.toggleLayout(localStorage.getItem('layout') || themeConfig.layout));
-        dispatch(themeConfigSlice.actions.toggleRTL(localStorage.getItem('rtlClass') || themeConfig.rtlClass));
-        dispatch(themeConfigSlice.actions.toggleAnimation(localStorage.getItem('animation') || themeConfig.animation));
-        dispatch(themeConfigSlice.actions.toggleNavbar(localStorage.getItem('navbar') || themeConfig.navbar));
-        dispatch(themeConfigSlice.actions.toggleSemidark(localStorage.getItem('semidark') || themeConfig.semidark));
+        dispatch(toggleTheme(localStorage.getItem('theme') || themeConfig.theme));
+        dispatch(toggleMenu(localStorage.getItem('menu') || themeConfig.menu));
+        dispatch(toggleLayout(localStorage.getItem('layout') || themeConfig.layout));
+        dispatch(toggleRTL(localStorage.getItem('rtlClass') || themeConfig.rtlClass));
+        dispatch(toggleAnimation(localStorage.getItem('animation') || themeConfig.animation));
+        dispatch(toggleNavbar(localStorage.getItem('navbar') || themeConfig.navbar));
+        dispatch(toggleSemidark(localStorage.getItem('semidark') || themeConfig.semidark));
         // locale
