+++ lib/createAppSlice.ts
@@ -1,0 +1,6 @@
+import { asyncThunkCreator, buildCreateSlice } from "@reduxjs/toolkit";
+
+// `buildCreateSlice` allows us to create a slice with async thunks.
+export const createAppSlice = buildCreateSlice({
+  creators: { asyncThunk: asyncThunkCreator },
+});
